Atualização das Class no projeto ClienteEspecial

+++ SquadBank/src/conta/Conta.java
@@ -4,10 +4,10 @@ public class Conta {
 
 	private String cliente;
 	private String cpf;
-	private double agencia;
+	private int agencia;
 	private double saldo;
 
-	public Conta(String cliente, String cpf, double agencia, double saldo) {
+	public Conta(String cliente, String cpf, int agencia, double saldo) {
 		this.cliente = cliente;
 		this.cpf = cpf;
 		this.agencia = agencia;
